Upgrade to VS2019 project and one more code cleanup.

+++ TMG/g_combat.c
@@ -268,9 +268,9 @@ static int CheckArmor (edict_t *ent, vec3_t point, vec3_t normal, int damage, in
 	armor = GetItemByIndex (index);
 
 	if (dflags & DAMAGE_ENERGY)
-		save = ceil(((gitem_armor_t *)armor->info)->energy_protection*damage);
+		save = ceilf(((gitem_armor_t *)armor->info)->energy_protection * damage);
 	else
-		save = ceil(((gitem_armor_t *)armor->info)->normal_protection*damage);
+		save = ceilf(((gitem_armor_t *)armor->info)->normal_protection * damage);
 	if (save >= client->pers.inventory[index])
 		save = client->pers.inventory[index];
